UHF-8525: Used unused debug method, moved test to correct namespace
